requirements.txt
semanticscholar
python-dotenv
langchain
langchain_groq